Fix track layout by swapping a margin to padding
The top spacing was set as margin instead of padding, meaning that when users would tap the top half of a track, the listener associated with the view wouldn't trigger since that part of the view was part of the margin and not the actual view. Changing this to padding fixes that.